+++ common/src/main/java/club/iananderson/seasonhud/impl/seasons/Calendar.java
@@ -20,7 +20,7 @@ public class Calendar {
     if (Config.needCalendar.get() && calendarLoaded && player != null) {
       Inventory inv = player.getInventory();
       int slot = findCalendar(inv, Services.SEASON.calendar()) + Services.SEASON.findCuriosCalendar(player,
-                                                                                                    Services.SEASON.calendar());
+          Services.SEASON.calendar());
 
       return (slot >= 0);
     } else {