+++ common/src/main/java/club/iananderson/seasonhud/impl/seasons/SeasonList.java
@@ -9,33 +9,22 @@ import net.minecraftforge.common.ForgeConfigSpec;
 
 public enum SeasonList {
   SPRING(0, Component.translatable("desc.seasonhud.spring"), "spring", "\uEA00", Config.springColor,
-         Config.springColor.getDefault(), Config.getSpringColor(), Rgb.seasonMap(Config.getSpringColor())), SUMMER(1,
-                                                                                                                   Component.translatable(
-                                                                                                                       "desc.seasonhud.summer"),
-                                                                                                                   "summer",
-                                                                                                                   "\uEA01",
-                                                                                                                   Config.summerColor,
-                                                                                                                   Config.summerColor.getDefault(),
-                                                                                                                   Config.getSummerColor(),
-                                                                                                                   Rgb.seasonMap(
-                                                                                                                       Config.getSummerColor())), AUTUMN(
-      2, Component.translatable("desc.seasonhud.autumn"), "autumn", "\uEA02", Config.autumnColor,
-      Config.autumnColor.getDefault(), Config.getAutumnColor(), Rgb.seasonMap(Config.getAutumnColor())), WINTER(3,
-                                                                                                                Component.translatable(
-                                                                                                                    "desc.seasonhud.winter"),
-                                                                                                                "winter",
-                                                                                                                "\uEA03",
-                                                                                                                Config.winterColor,
-                                                                                                                Config.winterColor.getDefault(),
-                                                                                                                Config.getWinterColor(),
-                                                                                                                Rgb.seasonMap(
-                                                                                                                    Config.getWinterColor())), DRY(
-      4, Component.translatable("desc.seasonhud.dry"), "dry", "\uEA04", Config.dryColor, Config.dryColor.getDefault(),
-      Config.getDryColor(), Rgb.seasonMap(Config.getDryColor())), WET(5, Component.translatable("desc.seasonhud.wet"),
-                                                                      "wet", "\uEA05", Config.wetColor,
-                                                                      Config.wetColor.getDefault(),
-                                                                      Config.getWetColor(),
-                                                                      Rgb.seasonMap(Config.getWetColor()));
+      Config.springColor.getDefault(), Config.getSpringColor(), Rgb.seasonMap(Config.getSpringColor())),
+
+  SUMMER(1, Component.translatable("desc.seasonhud.summer"), "summer", "\uEA01", Config.summerColor,
+      Config.summerColor.getDefault(), Config.getSummerColor(), Rgb.seasonMap(Config.getSummerColor())),
+
+  AUTUMN(2, Component.translatable("desc.seasonhud.autumn"), "autumn", "\uEA02", Config.autumnColor,
+      Config.autumnColor.getDefault(), Config.getAutumnColor(), Rgb.seasonMap(Config.getAutumnColor())),
+
+  WINTER(3, Component.translatable("desc.seasonhud.winter"), "winter", "\uEA03", Config.winterColor,
+      Config.winterColor.getDefault(), Config.getWinterColor(), Rgb.seasonMap(Config.getWinterColor())),
+
+  DRY(4, Component.translatable("desc.seasonhud.dry"), "dry", "\uEA04", Config.dryColor, Config.dryColor.getDefault(),
+      Config.getDryColor(), Rgb.seasonMap(Config.getDryColor())),
+
+  WET(5, Component.translatable("desc.seasonhud.wet"), "wet", "\uEA05", Config.wetColor, Config.wetColor.getDefault(),
+      Config.getWetColor(), Rgb.seasonMap(Config.getWetColor()));
 
   public static EnumSet<SeasonList> seasons = EnumSet.allOf(SeasonList.class);
   private final int id;
@@ -48,8 +37,8 @@ public enum SeasonList {
   private int seasonColor;
 
   SeasonList(int id, Component seasonName, String fileName, String iconChar,
-             ForgeConfigSpec.ConfigValue<Integer> seasonColorConfig, int defaultColor, int seasonColor,
-             HashMap<String, Integer> rgbMap) {
+      ForgeConfigSpec.ConfigValue<Integer> seasonColorConfig, int defaultColor, int seasonColor,
+      HashMap<String, Integer> rgbMap) {
     this.id = id;
     this.seasonName = seasonName;
     this.seasonFileName = fileName;